Time: 4 ms (99.97%) | Memory: 47.4 MB (81.71%) - LeetSync

+++ 713-subarray-product-less-than-k/subarray-product-less-than-k.java
@@ -1,7 +1,7 @@
 class Solution {
     public int numSubarrayProductLessThanK(int[] nums, int k) {
         int ans = 0, s = 0, e = 0, n = nums.length;
-        long mult = 1;
+        int mult = 1;
         while (e < n) {
             mult *= nums[e++];
             while (s < e && mult >= k) {
